requirements.txt
streamlit==1.37.0
python-docx==1.1.2
pdfminer.six==20231228
